chore: deleted redundant files
